requirements.txt
pandas~=1.5.3
kaggle~=1.5.12
prefect~=2.8.1
tqdm~=4.64.1
omegaconf~=2.3.0
matplotlib~=3.6.3
datasets~=2.9.0
transformers~=4.26.1
typing==3.7.4.3
pillow~=9.4.0
torchvision~=0.12.0
setuptools~=67.1.0